+++ TDDBowling/Exceptions/MaxPlayerReachedException.cs
@@ -5,7 +5,7 @@ namespace TDDBowling
     public class MaxPlayerReachedException : Exception
     {
         public MaxPlayerReachedException()
-        : base("A game can contain only 8 players")
+			: base("A game can contain only 8 players")
         {
             
         }
