+++ TDDBowling/Exceptions/MaxTwoShootsException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace TDDBowling
+{
+    public class MaxTwoShootsException : Exception
+    {
+		public MaxTwoShootsException()
+			: base("A player can't shoot more than 2 times per turn.")
+		{
+
+		}
+    }
+}
